LPS-55532 Normalize the path of template in URLResourceParser, then all the subclasses of URLResourceParser would not need to worry about path problem.
