+++ portal-impl/src/com/liferay/portal/template/ClassLoaderResourceParser.java
@@ -17,14 +17,8 @@ package com.liferay.portal.template;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.template.TemplateConstants;
-import com.liferay.portal.kernel.util.CharPool;
-import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.kernel.util.StringUtil;
 
 import java.net.URL;
-
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * @author Tina Tian
@@ -56,78 +50,7 @@ public class ClassLoaderResourceParser extends URLResourceParser {
 			_log.debug("Loading " + templateId);
 		}
 
-		templateId = normalizePath(templateId);
-
 		return _classLoader.getResource(templateId);
-	}
-
-	protected static String normalizePath(String path) {
-		List<String> elements = new ArrayList<>();
-
-		boolean absolutePath = false;
-
-		int previousIndex = -1;
-
-		for (int index;
-			 (index = path.indexOf(CharPool.SLASH, previousIndex + 1)) != -1;
-			 previousIndex = index) {
-
-			if ((previousIndex + 1) == index) {
-
-				// Starts with "/"
-
-				if (previousIndex == -1) {
-					absolutePath = true;
-
-					continue;
-				}
-
-				// "//" is illegal
-
-				throw new IllegalArgumentException(
-					"Unable to parse path " + path);
-			}
-
-			String pathElement = path.substring(previousIndex + 1, index);
-
-			// "." needs no handling
-
-			if (pathElement.equals(StringPool.PERIOD)) {
-				continue;
-			}
-
-			// ".." pops up stack
-
-			if (pathElement.equals(StringPool.DOUBLE_PERIOD)) {
-				if (elements.isEmpty()) {
-					throw new IllegalArgumentException(
-						"Unable to parse path " + path);
-				}
-
-				elements.remove(elements.size() - 1);
-
-				continue;
-			}
-
-			// Others push down stack
-
-			elements.add(pathElement);
-		}
-
-		if (previousIndex == -1) {
-			elements.add(path);
-		}
-		else if ((previousIndex + 1) < path.length()) {
-			elements.add(path.substring(previousIndex + 1));
-		}
-
-		String normalizedPath = StringUtil.merge(elements, StringPool.SLASH);
-
-		if (absolutePath) {
-			normalizedPath = StringPool.SLASH.concat(normalizedPath);
-		}
-
-		return normalizedPath;
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
