+++ portal-impl/src/com/liferay/portal/template/URLResourceParser.java
@@ -17,10 +17,16 @@ package com.liferay.portal.template;
 import com.liferay.portal.kernel.template.TemplateException;
 import com.liferay.portal.kernel.template.TemplateResource;
 import com.liferay.portal.kernel.template.URLTemplateResource;
+import com.liferay.portal.kernel.util.CharPool;
+import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.kernel.util.StringUtil;
 
 import java.io.IOException;
 
 import java.net.URL;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * @author Tina Tian
@@ -30,6 +36,8 @@ public abstract class URLResourceParser implements TemplateResourceParser {
 	@Override
 	public TemplateResource getTemplateResource(String templateId)
 		throws TemplateException {
+
+		templateId = normalizePath(templateId);
 
 		try {
 			URL url = getURL(templateId);
@@ -46,5 +54,74 @@ public abstract class URLResourceParser implements TemplateResourceParser {
 	}
 
 	public abstract URL getURL(String templateId) throws IOException;
+
+	protected static String normalizePath(String path) {
+		List<String> elements = new ArrayList<>();
+
+		boolean absolutePath = false;
+
+		int previousIndex = -1;
+
+		for (int index;
+			 (index = path.indexOf(CharPool.SLASH, previousIndex + 1)) != -1;
+			 previousIndex = index) {
+
+			if ((previousIndex + 1) == index) {
+
+				// Starts with "/"
+
+				if (previousIndex == -1) {
+					absolutePath = true;
+
+					continue;
+				}
+
+				// "//" is illegal
+
+				throw new IllegalArgumentException(
+					"Unable to parse path " + path);
+			}
+
+			String pathElement = path.substring(previousIndex + 1, index);
+
+			// "." needs no handling
+
+			if (pathElement.equals(StringPool.PERIOD)) {
+				continue;
+			}
+
+			// ".." pops up stack
+
+			if (pathElement.equals(StringPool.DOUBLE_PERIOD)) {
+				if (elements.isEmpty()) {
+					throw new IllegalArgumentException(
+						"Unable to parse path " + path);
+				}
+
+				elements.remove(elements.size() - 1);
+
+				continue;
+			}
+
+			// Others push down stack
+
+			elements.add(pathElement);
+		}
+
+		if (previousIndex == -1) {
+			elements.add(path);
+		}
+		else if ((previousIndex + 1) < path.length()) {
+			elements.add(path.substring(previousIndex + 1));
+		}
+
+		String normalizedPath = StringUtil.merge(elements, StringPool.SLASH);
+
+		if (absolutePath) {
+			normalizedPath = StringPool.SLASH.concat(normalizedPath);
+		}
+
+		return normalizedPath;
+	}
 
 }
